+++ src/main/kotlin/ara/analysis/LivenessAnalysis.kt
@@ -187,7 +187,7 @@ class LivenessAnalysis(val program: Syntax.Program) : Analysis<Unit>() {
 
         private fun verifyUses(instruction: Syntax.Instruction) {
             when (instruction) {
-                is Syntax.Assignment -> {
+                is Syntax.ArithmeticAssignment -> {
                     if (instruction.arithmetic != null) {
                         val resources = instruction.arithmetic.value.asResourcePaths().toSet()
                         for (resource in resources) {