+++ src/main/kotlin/ara/analysis/TypeDefinitionAnalysis.kt
@@ -5,6 +5,9 @@ import ara.syntax.Syntax
 import ara.types.TypeUnification
 import ara.types.TypeUnification.unify
 
+/**
+ * Analysis pass collecting all user-defined types.
+ */
 class TypeDefinitionAnalysis(private val program: Syntax.Program) : Analysis<Unit>() {
 
     override fun runAnalysis() {